pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "event-management-system"
version = "0.2.0"
description = "Revolutionary Event Management Platform with AI-powered features"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "cryptography>=41.0.0",
    "stripe>=5.5.0",
    "requests>=2.31.0",
    "python-socketio>=5.8.0",
    "flask-socketio>=5.3.0",
    "celery>=5.3.0",
    "redis>=4.6.0",
    "Pillow>=10.0.0",
    "qrcode>=7.4.2",
    "reportlab>=4.0.4",
    "google-auth>=2.22.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.97.0",
    "msal>=1.24.0",
    "tweepy>=4.14.0",
    "slack-sdk>=3.21.0",
    "PyGithub>=1.59.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0"
]

[tool.setuptools]
py-modules = ["app", "models", "routes", "ai_engine", "collaboration", "websocket_handlers", "virtual_events", "analytics", "integrations"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["templates/*", "static/*", "static/**/*"]
